Alternate solution to async image cell reuse issue
Previously, when item.posterPath was captured as posterPath,
it would always be the path of the item irrespective of cell
changes so the comparison in the callback is always true.

This change adds a property to the cell, which is written to
each time the cell is reused, so when the callback happens it
checks the path for the cells current content.